Complete rechecking full plugin

+++ includes/Admin/Admin.php
@@ -2,7 +2,9 @@
 
 namespace UTMManager\Admin;
 
-defined( 'ABSPATH' ) || exit;
+use UTMManager\Admin\ListTables\LeadsListTable;
+
+defined( 'ABSPATH' ) || exit; // Exit if accessed directly.
 
 /**
  * Class Admin.
@@ -13,7 +15,9 @@ defined( 'ABSPATH' ) || exit;
 class Admin {
 
 	/**
-	 * Constructor.
+	 * Admin constructor.
+	 *
+	 * @since 1.0.0
 	 */
 	public function __construct() {
 		add_action( 'admin_menu', array( $this, 'add_menu' ) );
@@ -24,22 +28,25 @@ class Admin {
 
 	/**
 	 * Add menu page.
+	 *
+	 * @since 1.0.0
+	 * @return void
 	 */
 	public function add_menu() {
 		add_menu_page(
-			'UTM Manager',
-			'UTM Manager',
+			__( 'UTM Manager', 'utm-manager' ),
+			__( 'UTM Manager', 'utm-manager' ),
 			'manage_options',
 			'utm-manager',
 			null,
 			'dashicons-admin-links',
-			'100',
+			'55.9',
 		);
 
 		$load = add_submenu_page(
 			'utm-manager',
-			'Leads',
-			'Leads',
+			__( 'Leads', 'utm-manager' ),
+			__( 'Leads', 'utm-manager' ),
 			'manage_options',
 			'utm-manager',
 			array( $this, 'render_page' ),
@@ -91,11 +98,14 @@ class Admin {
 	}
 
 	/**
-	 * Render leads menu.
+	 * Render leads page.
+	 *
+	 * @since 1.0.0
+	 * @return void
 	 */
 	public function render_page() {
 		wp_verify_nonce( '_nonce' );
-		$view = isset( $_GET['view'] ) ? absint( $_GET['view'] ) : 0;
+		$view = isset( $_GET['view'] ) ? absint( wp_unslash( $_GET['view'] ) ) : 0;
 
 		if ( $view ) {
 			$lead = utmm_get_lead( $view );
@@ -104,35 +114,42 @@ class Admin {
 				wp_safe_redirect( remove_query_arg( 'view' ) );
 				exit();
 			}
-		}
 
-		if ( $view ) {
 			include __DIR__ . '/views/view-lead.php';
 		} else {
-			$list_table = new \UTMManager\Admin\ListTables\LeadsListTable();
+			$list_table = new LeadsListTable();
 			$list_table->prepare_items();
 			include __DIR__ . '/views/leads.php';
 		}
 	}
 
 	/**
-	 * Handle leads list table.
+	 * Handle list table actions.
+	 *
+	 * @since 1.0.0
+	 * @return void
 	 */
 	public function handle_list_table_actions() {
-		$wp_list_table = new \UTMManager\Admin\ListTables\LeadsListTable();
+
+		if ( ! current_user_can( 'manage_options' ) ) {
+			utm_manager()->add_flash_notice( esc_html__( 'You do not have permission to perform this action.', 'utm-manager' ), 'error' );
+			$redirect_url = remove_query_arg( array( 'action', 'ids', '_wpnonce' ) );
+			wp_safe_redirect( $redirect_url );
+			exit;
+		}
+
+		$wp_list_table = new LeadsListTable();
 		$wp_list_table->process_bulk_action();
 
 		if ( 'delete' === $wp_list_table->current_action() ) {
-			// Verify nonce.
 			check_admin_referer( 'bulk-leads' );
 
-			// Get selected item IDs.
-			$item_ids = isset( $_REQUEST['ids'] ) ? (array) $_REQUEST['ids'] : array();
-
-			// Process delete action.
+			$ids       = isset( $_GET['ids'] ) ? map_deep( wp_unslash( $_GET['ids'] ), 'intval' ) : array();
+			$ids       = wp_parse_id_list( $ids );
 			$performed = 0;
-			foreach ( $item_ids as $item_id ) {
-				$lead = utmm_get_lead( $item_id );
+
+			foreach ( $ids as $id ) {
+				$lead = utmm_get_lead( $id );
 				if ( $lead && wp_delete_post( $lead->ID, true ) ) {
 					++$performed;
 				}
@@ -140,12 +157,12 @@ class Admin {
 
 			if ( ! empty( $performed ) ) {
 				// translators: %s: number of accounts.
-				utm_manager()->add_flash_notice( sprintf( __( '%s item(s) deleted successfully.', 'utm-manager' ), number_format_i18n( $performed ) ), 'success' );
+				utm_manager()->add_flash_notice( sprintf( esc_html__( '%s item(s) deleted successfully.', 'utm-manager' ), number_format_i18n( $performed ) ) );
 			}
 
 			if ( ! headers_sent() ) {
 				// Redirect to avoid resubmission.
-				$redirect_url = remove_query_arg( array( 'action', 'custom_item', '_wpnonce' ) );
+				$redirect_url = remove_query_arg( array( 'action', 'ids', '_wpnonce' ) );
 				wp_safe_redirect( $redirect_url );
 				exit;
 			}
@@ -158,6 +175,7 @@ class Admin {
 	 * @param string $hook Hook name.
 	 *
 	 * @since 1.0.0
+	 * @return void
 	 */
 	public function enqueue_scripts( $hook ) {
 		$screens = array(
@@ -165,7 +183,7 @@ class Admin {
 			'utm-manager_page_utmm-settings',
 		);
 
-		wp_register_style( 'utmm-admin', UTMM_URL . 'assets/dist/css/utmm-admin.css', array(), '1.0.0' );
+		wp_register_style( 'utmm-admin', UTMM_ASSETS_URL . 'dist/css/utmm-admin.css', array(), UTMM_VERSION );
 
 		if ( in_array( $hook, $screens, true ) ) {
 			wp_enqueue_style( 'utmm-admin' );
